+++ bench/index.js
@@ -39,6 +39,7 @@ import kleur from 'kleur';
 import * as kleurColors from 'kleur/colors';
 import picocolors from 'picocolors';
 import { Ansis, green, red, yellow, hex } from 'ansis';
+//import { Ansis, green, red, yellow, hex } from '../_ignored/index-2.0.3.min.js';
 
 const log = console.log;