+++ test/flags.test.js
@@ -10,79 +10,71 @@ const TEST_PATH = path.resolve('./test/');
 //   - test NO_COLOR=1
 
 describe('enable colors', () => {
-  test(`--color`, (done) => {
+  test(`--color`, () => {
     const filename = path.join(TEST_PATH, './cli/output.js');
     const received = execScriptSync(filename, ['--color']);
     const expected =
       '\x1b[31mred\x1b[39m|\x1b[38;2;80;80;80mrgb\x1b[39m|\x1b[48;2;80;80;80mbgRgb\x1b[49m|\x1b[38;2;255;255;255mhex\x1b[39m|\x1b[48;2;255;255;255mbgHex\x1b[49m';
     expect(esc(received)).toEqual(esc(expected));
-    done();
   });
 
-  test(`--color=true`, (done) => {
+  test(`--color=true`, () => {
     const filename = path.join(TEST_PATH, './cli/output.js');
     const received = execScriptSync(filename, ['--color=true']);
     const expected =
       '\x1b[31mred\x1b[39m|\x1b[38;2;80;80;80mrgb\x1b[39m|\x1b[48;2;80;80;80mbgRgb\x1b[49m|\x1b[38;2;255;255;255mhex\x1b[39m|\x1b[48;2;255;255;255mbgHex\x1b[49m';
     expect(esc(received)).toEqual(esc(expected));
-    done();
   });
 
-  test(`--color=always`, (done) => {
+  test(`--color=always`, () => {
     const filename = path.join(TEST_PATH, './cli/output.js');
     const received = execScriptSync(filename, ['--color=always']);
     const expected =
       '\x1b[31mred\x1b[39m|\x1b[38;2;80;80;80mrgb\x1b[39m|\x1b[48;2;80;80;80mbgRgb\x1b[49m|\x1b[38;2;255;255;255mhex\x1b[39m|\x1b[48;2;255;255;255mbgHex\x1b[49m';
     expect(esc(received)).toEqual(esc(expected));
-    done();
   });
 
-  test(`FORCE_COLOR=true`, (done) => {
+  test(`FORCE_COLOR=true`, () => {
     const filename = path.join(TEST_PATH, './cli/output.js');
     const received = execScriptSync(filename, [], ['FORCE_COLOR=true']);
     const expected =
       '\x1b[31mred\x1b[39m|\x1b[38;2;80;80;80mrgb\x1b[39m|\x1b[48;2;80;80;80mbgRgb\x1b[49m|\x1b[38;2;255;255;255mhex\x1b[39m|\x1b[48;2;255;255;255mbgHex\x1b[49m';
 
     expect(esc(received)).toEqual(esc(expected));
-    done();
   });
 
-  test(`FORCE_COLOR=1`, (done) => {
+  test(`FORCE_COLOR=1`, () => {
     const filename = path.join(TEST_PATH, './cli/output.js');
     const received = execScriptSync(filename, [], ['FORCE_COLOR=1']);
     const expected =
       '\x1b[31mred\x1b[39m|\x1b[38;2;80;80;80mrgb\x1b[39m|\x1b[48;2;80;80;80mbgRgb\x1b[49m|\x1b[38;2;255;255;255mhex\x1b[39m|\x1b[48;2;255;255;255mbgHex\x1b[49m';
 
     expect(esc(received)).toEqual(esc(expected));
-    done();
   });
 });
 
 describe('disable colors', () => {
-  test(`--no-color`, (done) => {
+  test(`--no-color`, () => {
     const filename = path.join(TEST_PATH, './cli/output.js');
     // flags has priority over env variable
     const received = execScriptSync(filename, ['--no-color'], ['FORCE_COLOR=1']);
     const expected = 'red|rgb|bgRgb|hex|bgHex';
     expect(esc(received)).toEqual(esc(expected));
-    done();
   });
 
-  test(`--color=false`, (done) => {
+  test(`--color=false`, () => {
     const filename = path.join(TEST_PATH, './cli/output.js');
     // flags has priority over env variable
     const received = execScriptSync(filename, ['--color=false'], ['FORCE_COLOR=1']);
     const expected = 'red|rgb|bgRgb|hex|bgHex';
     expect(esc(received)).toEqual(esc(expected));
-    done();
   });
 
-  test(`--color=never`, (done) => {
+  test(`--color=never`, () => {
     const filename = path.join(TEST_PATH, './cli/output.js');
     // flags has priority over env variable
     const received = execScriptSync(filename, ['--color=never'], ['FORCE_COLOR=1']);
     const expected = 'red|rgb|bgRgb|hex|bgHex';
     expect(esc(received)).toEqual(esc(expected));
-    done();
   });
 });